+++ xphyle/__init__.py
@@ -124,14 +124,18 @@ class FileLikeWrapper(EventManager, FileLikeBase):
     Args:
         fileobj: The file-like object to wrap.
         compression: Whether the wrapped file is compressed.
+        close_fileobj: Whether to close the wrapped file object when closing
+            this wrapper.
+
     """
     def __init__(
-            self, fileobj: FileLike, compression: CompressionArg = False
-            ) -> None:
+            self, fileobj: FileLike, compression: CompressionArg = False,
+            close_fileobj: bool = True) -> None:
         EventManager.__init__(self)
         self._fileobj = fileobj
         self._iterator: Iterator = None
         self.compression = compression
+        self.close_fileobj = close_fileobj
 
     def __next__(self) -> bytes:
         return next(iter(self))
@@ -204,7 +208,8 @@ class FileLikeWrapper(EventManager, FileLikeBase):
         self._fire_listeners(EventType.CLOSE)
 
     def _close(self) -> None:
-        self._fileobj.close()
+        if self.close_fileobj:
+            self._fileobj.close()
 
     # Pass-through methods
 
@@ -277,7 +282,8 @@ class FileWrapper(FileLikeWrapper):
     def __init__(
             self, source: PathOrFile, mode: ModeArg = 'w',
             compression: CompressionArg = False,
-            name: Union[str, PurePath] = None, **kwargs) -> None:
+            name: Union[str, PurePath] = None, close_fileobj: bool = True,
+            **kwargs) -> None:
         if isinstance(source, Path):
             self._path = source
             source_fileobj = xopen(
@@ -287,7 +293,8 @@ class FileWrapper(FileLikeWrapper):
             if name is None and hasattr(source_fileobj, 'name'):
                 name = str(getattr(source_fileobj, 'name'))
             self._path = Path(name) if name else None
-        super().__init__(source_fileobj, compression=compression)
+        super().__init__(
+            source_fileobj, compression=compression, close_fileobj=close_fileobj)
         self._name = str(name)
         if mode is None and hasattr(source, 'mode'):
             self._mode = getattr(source_fileobj, 'mode')
@@ -309,14 +316,6 @@ class FileWrapper(FileLikeWrapper):
         return getattr(self, '_path', None)
 
 
-class OpenFileWrapper(FileWrapper):
-    """Wrapper around a file that is already open. Closing the wrapper does
-    not close the underlying file.
-    """
-    def _close(self) -> None:
-        self._fileobj.flush()
-
-
 class BufferWrapper(FileWrapper):
     """Wrapper around a string/bytes buffer.
 
@@ -324,11 +323,13 @@ class BufferWrapper(FileWrapper):
         fileobj: The fileobj to wrap (the raw or wrapped buffer).
         buffer: The raw buffer.
         compression: Compression type.
+        close_fileobj: Whether to close the buffer when closing this wrapper.
     """
     def __init__(
             self, fileobj: PathOrFile, buffer: Union[io.StringIO, io.BytesIO],
-            compression: CompressionArg = False, name: str = None) -> None:
-        super().__init__(fileobj, compression=compression, name=name)
+            compression: CompressionArg = False, name: str = None, **kwargs
+            ) -> None:
+        super().__init__(fileobj, compression=compression, name=name, **kwargs)
         self.buffer = buffer
 
     def getvalue(self) -> AnyChar:
@@ -342,11 +343,10 @@ class BufferWrapper(FileWrapper):
     def _close(self):
         if self.compression:
             self._fileobj.close()
-            value = self.getvalue()
-        else:
-            value = self.getvalue()
+            setattr(self, '_value', self.buffer.getvalue())
+        elif self.close_fileobj:
+            setattr(self, '_value', self.buffer.getvalue())
             self._fileobj.close()
-        setattr(self, '_value', value)
 
 
 class StdWrapper(FileLikeWrapper):
@@ -788,8 +788,7 @@ def open_(
                 raise ValueError(
                     "'wrap_fileobj must be True if 'path' is not file-like")
         else:
-            if not is_fileobj:
-                kwargs['context_wrapper'] = True
+            kwargs['context_wrapper'] = True
             try:
                 with xopen(target, mode, **kwargs) as fileobj:
                     yield fileobj
@@ -908,7 +907,7 @@ def xopen(
         if not mode:
             mode = FileMode(access='r', coding='t' if is_str else 'b')
         is_buffer = True
-    elif ((is_str or is_path) == (file_type is FileType.FILELIKE) or
+    elif ((is_str or is_path or is_buffer) == (file_type is FileType.FILELIKE) or
           is_std != (file_type is FileType.STDIO) or
           is_buffer != (file_type is FileType.BUFFER)):
         raise ValueError(
@@ -1154,13 +1153,13 @@ def xopen(
         if is_std:
             fileobj = StdWrapper(fileobj, compression=compression)
         elif file_type == FileType.BUFFER:
-            fileobj = BufferWrapper(fileobj, buffer, compression=compression)
+            fileobj = BufferWrapper(
+                fileobj, buffer, compression=compression,
+                close_fileobj=close_fileobj)
         else:
-            file_wrapper_class = FileWrapper
-            if file_type == FileType.FILELIKE and not close_fileobj:
-                file_wrapper_class = OpenFileWrapper
-            fileobj = file_wrapper_class(
-                fileobj, name=name, mode=mode, compression=compression)
+            fileobj = FileWrapper(
+                fileobj, name=name, mode=mode, compression=compression,
+                close_fileobj=close_fileobj)
 
     return fileobj